refactor: split wizard probe helpers
